scripts/requirements.txt
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.6.15
charset-normalizer==3.4.2
idna==3.10
psycopg2-binary==2.9.10
python-dotenv==1.1.1
requests==2.32.4
soupsieve==2.7
typing_extensions==4.14.0
urllib3==2.5.0
